requirements.txt
Django==2.2.10
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-decouple==3.3
dj-database-url==0.5.0
boto3==1.12.0
gunicorn==20.0.4
django-solo==1.1.3
